+++ src/rest/decorators/disableInterceptor.decorator.ts
@@ -2,7 +2,8 @@ import {Type} from '@angular/core';
 import {isBlank} from '@jscrpt/common';
 
 import {RESTClient} from '../common';
-import {RestDisabledInterceptors} from '../rest.interface';
+import {RestDisabledInterceptors, RestMethodMiddlewares} from '../rest.interface';
+import {IgnoredInterceptorsMiddleware} from '../middlewares';
 
 /**
  * Disables specified type of http client interceptor for all calls of applied method
@@ -10,7 +11,8 @@ import {RestDisabledInterceptors} from '../rest.interface';
  */
 export function DisableInterceptor<TType>(interceptorType: Type<TType>)
 {
-    return function(_target: RESTClient, _propertyKey: string, descriptor: RestDisabledInterceptors<TType>)
+    return function(_target: RESTClient, _propertyKey: string, descriptor: RestDisabledInterceptors<TType> &
+                                                                           RestMethodMiddlewares)
     {
         if(isBlank(interceptorType))
         {
@@ -22,6 +24,7 @@ export function DisableInterceptor<TType>(interceptorType: Type<TType>)
             descriptor.disabledInterceptors = [];
         }
 
+        descriptor.middlewareTypes.push(IgnoredInterceptorsMiddleware);
         descriptor.disabledInterceptors.push(interceptorType);
 
         return descriptor;