+++ src/rest/decorators/progressIndicatorGroup.decorator.ts
@@ -1,7 +1,8 @@
 import {LocalProgressIndicatorName} from '@anglr/common';
 
 import {RESTClient} from '../common';
-import {AdditionalInfoPropertyDescriptor} from '../rest.interface';
+import {AdditionalInfoPropertyDescriptor, RestMethodMiddlewares} from '../rest.interface';
+import {AdditionalDataMiddleware} from '../middlewares';
 
 /**
  * Allows to specify progress indicator group for displaying local progress indicator
@@ -9,13 +10,15 @@ import {AdditionalInfoPropertyDescriptor} from '../rest.interface';
  */
 export function ProgressIndicatorGroup(name: string)
 {
-    return function(_target: RESTClient, _propertyKey: string, descriptor: AdditionalInfoPropertyDescriptor<LocalProgressIndicatorName>)
+    return function(_target: RESTClient, _propertyKey: string, descriptor: AdditionalInfoPropertyDescriptor<LocalProgressIndicatorName> &
+                                                                           RestMethodMiddlewares)
     {
         if(!descriptor.additionalInfo)
         {
             descriptor.additionalInfo = {};
         }
 
+        descriptor.middlewareTypes.push(AdditionalDataMiddleware);
         descriptor.additionalInfo.progressGroupName = name;
 
         return descriptor;