+++ src/rest/defaults.ts
@@ -1,16 +1,18 @@
 import {Type} from '@angular/core';
 
 import {RestMiddleware} from './rest.interface';
-import {ReportProgressMiddleware, ResponseTypeMiddleware, ResponseTransformMiddleware, IgnoredInterceptorsMiddleware} from './middlewares';
+import {ReportProgressMiddleware, ResponseTypeMiddleware, IgnoredInterceptorsMiddleware, ResponseTransformMiddleware, AdditionalDataMiddleware, ProducesMiddleware} from './middlewares';
 
 /**
  * Definition of basic default array of rest middlewares order
  */
 export const BASIC_DEFAULT_REST_MIDDLEWARES_ORDER: Type<RestMiddleware>[] =
 [
+    ProducesMiddleware,
+    AdditionalDataMiddleware,
+    IgnoredInterceptorsMiddleware,
     ResponseTransformMiddleware,
     ResponseTypeMiddleware,
-    IgnoredInterceptorsMiddleware,
     ReportProgressMiddleware
 ];
 
@@ -19,8 +21,6 @@ export const BASIC_DEFAULT_REST_MIDDLEWARES_ORDER: Type<RestMiddleware>[] =
  */
 export const BASIC_DEFAULT_REST_METHOD_MIDDLEWARES: Type<RestMiddleware>[] =
 [
-    ResponseTransformMiddleware,
     ResponseTypeMiddleware,
-    IgnoredInterceptorsMiddleware,
     ReportProgressMiddleware
 ];