+++ src/rest/middlewares/index.ts
@@ -2,3 +2,5 @@ export * from './reportProgress.middleware';
 export * from './responseType.middleware';
 export * from './responseTransform.middleware';
 export * from './ignoredInterceptors.middleware';
+export * from './additionalData.middleware';
+export * from './produces.middleware';